fix(middleware): scheme check should be performed before "home" redirect
The issue in #802 is caused by the redirect to "home" that is done before the $auth.check call in middleware.ts. The redirect causes the headers to be sent. The following $auth.check() call tries to set the cookies on the redirect response that was already sent, triggering the Exception.

Fixes #802